Fix typos in documentation
There were a couple typos in the documentation for the area chart that cause the graph not to render. A misspelling of height and an object that was missing the extra set of {}.

The changes that I've made reflect the changes made on this pull request (https://github.com/esbullington/react-d3/pull/273). They should be made here and on the gh-pages branch so it is up to date.